pyproject.toml
[tool.poetry]
name = "portfolio-website"
version = "0.1.0"
description = ""
authors = ["Fabrice Hafashimana <tumbafabruce@gmail.com>"]
readme = "README.md"
packages = [{include = "portfolio_website"}]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.3"
django-allauth = "^0.54.0"
django-decouple = "^2.1"
dj-database-url = "^2.0.0"
django-widget-tweaks = "^1.4.12"
django-simple-gravatar = "^1.0.1"
djangorestframework = "^3.14.0"
django-autoslug = "^1.9.9"
django-hitcount = "^1.3.5"
django-imagekit = "^4.1.0"
cloudinary = "^1.33.0"
django-ckeditor = "^6.6.1"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
